pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Aleksandr <gleksandr.myashin@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "^3.12.4"
Markdown = "^3.3.4"
Django = "^3.2.6"
django-filter = "^2.4.0"
aiogram = "^2.14.3"
parse = "^1.19.0"
Telethon = "^1.23.0"
aiohttp = "^3.7.4"
emoji = "^1.4.2"
loguru = "^0.5.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
